refactor(Fields JS): moved all of the input fields into the input folder and moved all other fields into the fields folder
