+++ ui/src/main/js/LoginPage.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import PasswordInput from 'field/input/PasswordInput';
 import TextInput from 'field/input/TextInput';
-import SubmitButton from 'field/input/SubmitButton';
+import SubmitButton from 'field/SubmitButton';
 import Header from 'component/common/Header';
 import { login } from 'store/actions/session';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
